refactor for be more easier to populate the database after the application has started

+++ src/main/java/br/com/mb/moviesbattleapp/domain/ranking/RankingResponse.java
@@ -4,11 +4,13 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
 public class RankingResponse {
     private Integer position;
-    private Integer points;
+    private BigDecimal points;
     private String player;
 }
